requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
dynaconf==3.2.3
exceptiongroup==1.1.3
Faker==19.8.1
idna==3.4
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
pytest-env==1.0.1
python-dateutil==2.8.2
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.4
six==1.16.0
tomli==2.0.1
urllib3==2.0.6
